+++ src/manage_pg.c
@@ -953,7 +953,7 @@ manage_create_sql_functions ()
            TASK_STATUS_STOP_REQUESTED_GIVEUP,
            TASK_STATUS_STOPPED,
            TASK_STATUS_INTERRUPTED,
-           TASK_STATUS_PENDING);
+           TASK_STATUS_QUEUED);
 
       sql ("CREATE OR REPLACE FUNCTION report_progress (integer)"
            " RETURNS integer AS $$"
@@ -1326,7 +1326,7 @@ manage_create_sql_functions ()
        "         WHEN $1 = %i"
        "         THEN 'Stopped'"
        "         WHEN $1 = %i"
-       "         THEN 'Pending'"
+       "         THEN 'Queued'"
        "         ELSE 'Interrupted'"
        "         END;"
        "$$ LANGUAGE SQL"
@@ -1343,7 +1343,7 @@ manage_create_sql_functions ()
        TASK_STATUS_STOP_REQUESTED,
        TASK_STATUS_STOP_WAITING,
        TASK_STATUS_STOPPED,
-       TASK_STATUS_PENDING);
+       TASK_STATUS_QUEUED);
 
   if (sql_int ("SELECT EXISTS (SELECT * FROM information_schema.tables"
                "               WHERE table_catalog = '%s'"