+++ src/manage_sql.c
@@ -2796,7 +2796,7 @@ keyword_applies_to_column (keyword_t *keyword, const char* column)
       && (strstr ("Done", keyword->string) == NULL)
       && (strstr ("New", keyword->string) == NULL)
       && (strstr ("Running", keyword->string) == NULL)
-      && (strstr ("Pending", keyword->string) == NULL)
+      && (strstr ("Queued", keyword->string) == NULL)
       && (strstr ("Stop Requested", keyword->string) == NULL)
       && (strstr ("Stopped", keyword->string) == NULL)
       && (strstr ("Interrupted", keyword->string) == NULL))
@@ -15226,7 +15226,7 @@ task_in_use (task_t task)
          || status == TASK_STATUS_DELETE_ULTIMATE_WAITING
          || status == TASK_STATUS_REQUESTED
          || status == TASK_STATUS_RUNNING
-         || status == TASK_STATUS_PENDING
+         || status == TASK_STATUS_QUEUED
          || status == TASK_STATUS_STOP_REQUESTED_GIVEUP
          || status == TASK_STATUS_STOP_REQUESTED
          || status == TASK_STATUS_STOP_WAITING;
@@ -16344,7 +16344,7 @@ stop_active_tasks ()
           case TASK_STATUS_DELETE_WAITING:
           case TASK_STATUS_REQUESTED:
           case TASK_STATUS_RUNNING:
-          case TASK_STATUS_PENDING:
+          case TASK_STATUS_QUEUED:
           case TASK_STATUS_STOP_REQUESTED_GIVEUP:
           case TASK_STATUS_STOP_REQUESTED:
           case TASK_STATUS_STOP_WAITING:
@@ -16389,7 +16389,7 @@ stop_active_tasks ()
        TASK_STATUS_DELETE_WAITING,
        TASK_STATUS_REQUESTED,
        TASK_STATUS_RUNNING,
-       TASK_STATUS_PENDING,
+       TASK_STATUS_QUEUED,
        TASK_STATUS_STOP_REQUESTED,
        TASK_STATUS_STOP_REQUESTED_GIVEUP,
        TASK_STATUS_STOP_WAITING);
@@ -17827,7 +17827,7 @@ set_task_requested (task_t task, task_status_t *status)
   run_status = task_run_status (task);
   if (run_status == TASK_STATUS_REQUESTED
       || run_status == TASK_STATUS_RUNNING
-      || run_status == TASK_STATUS_PENDING
+      || run_status == TASK_STATUS_QUEUED
       || run_status == TASK_STATUS_STOP_REQUESTED
       || run_status == TASK_STATUS_STOP_REQUESTED_GIVEUP
       || run_status == TASK_STATUS_STOP_WAITING
@@ -17902,7 +17902,7 @@ task_running_report (task_t task)
   task_status_t run_status = task_run_status (task);
   if (run_status == TASK_STATUS_REQUESTED
       || run_status == TASK_STATUS_RUNNING
-      || run_status == TASK_STATUS_PENDING)
+      || run_status == TASK_STATUS_QUEUED)
     {
       return (unsigned int) sql_int ("SELECT max(id) FROM reports"
                                      " WHERE task = %llu AND end_time IS NULL"
@@ -17910,7 +17910,7 @@ task_running_report (task_t task)
                                      " OR scan_run_status = %u);",
                                      task,
                                      TASK_STATUS_RUNNING,
-                                     TASK_STATUS_PENDING);
+                                     TASK_STATUS_QUEUED);
     }
   return (report_t) 0;
 }
@@ -17929,7 +17929,7 @@ task_iterator_current_report (iterator_t *iterator)
   task_status_t run_status = task_iterator_run_status (iterator);
   if (run_status == TASK_STATUS_REQUESTED
       || run_status == TASK_STATUS_RUNNING
-      || run_status == TASK_STATUS_PENDING
+      || run_status == TASK_STATUS_QUEUED
       || run_status == TASK_STATUS_DELETE_REQUESTED
       || run_status == TASK_STATUS_DELETE_ULTIMATE_REQUESTED
       || run_status == TASK_STATUS_STOP_REQUESTED
@@ -17951,7 +17951,7 @@ task_iterator_current_report (iterator_t *iterator)
                                      task,
                                      TASK_STATUS_REQUESTED,
                                      TASK_STATUS_RUNNING,
-                                     TASK_STATUS_PENDING,
+                                     TASK_STATUS_QUEUED,
                                      TASK_STATUS_DELETE_REQUESTED,
                                      TASK_STATUS_DELETE_ULTIMATE_REQUESTED,
                                      TASK_STATUS_STOP_REQUESTED,
@@ -24598,7 +24598,7 @@ delete_report_internal (report_t report)
                " OR scan_run_status = %u OR scan_run_status = %u);",
                report,
                TASK_STATUS_RUNNING,
-               TASK_STATUS_PENDING,
+               TASK_STATUS_QUEUED,
                TASK_STATUS_REQUESTED,
                TASK_STATUS_DELETE_REQUESTED,
                TASK_STATUS_DELETE_ULTIMATE_REQUESTED,
@@ -41744,7 +41744,7 @@ task_schedule_iterator_stop_due (iterator_t* iterator)
 
       if (run_status == TASK_STATUS_RUNNING
           || run_status == TASK_STATUS_REQUESTED
-          || run_status == TASK_STATUS_PENDING)
+          || run_status == TASK_STATUS_QUEUED)
         {
           time_t now, start;
 
@@ -52393,7 +52393,7 @@ delete_user (const char *user_id_arg, const char *name_arg, int ultimate,
         case TASK_STATUS_DELETE_WAITING:
         case TASK_STATUS_REQUESTED:
         case TASK_STATUS_RUNNING:
-        case TASK_STATUS_PENDING:
+        case TASK_STATUS_QUEUED:
         case TASK_STATUS_STOP_REQUESTED_GIVEUP:
         case TASK_STATUS_STOP_REQUESTED:
         case TASK_STATUS_STOP_WAITING: